+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/archive/plan02/day06/p014/Solution017_01.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan02.day06.p014;
+package com.noob.algorithm.daily.archive.plan02.day06.p014;
 
 import java.util.ArrayList;
 import java.util.HashMap;
