+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/archive/plan02/day06/p016/Solution078_01.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan02.day06.p016;
+package com.noob.algorithm.daily.archive.plan02.day06.p016;
 
 import java.util.*;
