+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/archive/plan02/day06/p017/Solution046_02.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan02.day06.p017;
+package com.noob.algorithm.daily.archive.plan02.day06.p017;
 
 import java.util.ArrayList;
 import java.util.List;
